+++ Rendering/Primitives/Triangle.cs
@@ -9,7 +9,7 @@ public struct Triangle() : IPrimitiveData
         0.0f,  0.5f, 0.0f
     ];
 
-    public ushort[] Indices { get; } =
+    public uint[] Indices { get; } =
     [
         0, 1, 2
     ];
